refactor: address review — lean connector, drop @FEEL, fix secret hint
Addresses johnBgood's review round:
- Extract the HTTP/config/execution logic into AppIntegrationsExecutor so the
  connector is a thin SPI entry point (like the HTTP connector's HttpService),
  and move the path/env/header constants alongside the code that uses them.
- Remove @FEEL from the auth models — variables are evaluated in the engine for
  outbound connectors, so the annotation is unnecessary.
- Update the API-key secret hint to the current {{secrets.X}} format.
- Set engineVersion to ^8.9 to match the planned backport target.
